Change argument type from int to long so that it can be used with getB() to retrieve all samples in frequency order.

+++ de.tudarmstadt.ukp.dkpro.core.api.frequency-asl/src/main/java/de/tudarmstadt/ukp/dkpro/core/api/frequency/util/FrequencyDistribution.java
@@ -281,7 +281,7 @@ public class FrequencyDistribution<T>
      * Returns the n most frequent samples in the distribution.
      * The ordering within in a group of samples with the same frequency is undefined.
      */
-    public List<T> getMostFrequentSamples(int n) {
+    public List<T> getMostFrequentSamples(long n) {
 
         List<T> topSamples = new ArrayList<T>();
